refactor(core): extract StoryDependencyGraphService from ChangeDetectionService
ChangeDetectionService fused two independent responsibilities: the module
dependency graph (build a reverse index from story files, watch builder file
events, patch incrementally) and the status pipeline (git diff -> affected
stories -> status store). This splits the graph half into a standalone
StoryDependencyGraphService that ChangeDetectionService now composes.

The two halves shared no state; the only coupling was a single
reverseIndex.lookup() call plus a patch-settle drain. The graph is now
independently constructible and testable, and "graph-only" usage is expressed
by constructing the tracker without a status publisher rather than via an
internal mode flag.

- StoryDependencyGraphService owns the adapter subscription, graph build,
  incremental patcher, patchQueue serialization, story-file reconciliation,
  and the OXC parse-pool teardown. It exposes lookup(), whenSettled() (a
  tail-snapshot patch-settle barrier), onStoryIndexInvalidated(), and
  onReady/onChange/onError/onUnavailable lifecycle callbacks. It does not
  touch the change-detection readiness signal.
- ChangeDetectionService keeps its public surface (start/onStoryIndexInvalidated/
  dispose) and stays the sole owner of git diffing, the status store, and
  readiness. dispose() is now idempotent and tears down the graph last.
- Shared story-file mapping moved to story-files.ts; shared test scaffolding
  moved to change-detection.test-helpers.ts.
- Graph-internal behavior is now tested directly in
  StoryDependencyGraphService.test.ts; ChangeDetectionService.test.ts retains
  the status / integration / readiness coverage.

No behavior change: dev-server wiring and the public barrel are untouched.
