+++ code/core/src/core-server/change-detection/ChangeDetectionService.ts
@@ -1,35 +1,24 @@
-import { writeFile } from 'node:fs/promises';
-
 import { join, normalize } from 'pathe';
 
 import { dequal } from 'dequal';
 import { logger } from 'storybook/internal/node-logger';
-import { disposeOxcParsePool } from 'storybook/internal/oxc-parser';
-import { getProjectRoot } from 'storybook/internal/common';
 import type {
   Presets,
-  StatusValue,
-  StoryIndex,
   Status,
   StatusStoreByTypeId,
+  StatusValue,
+  StoryIndex,
 } from 'storybook/internal/types';
 import { CHANGE_DETECTION_STATUS_TYPE_ID } from 'storybook/internal/types';
 
 import type { StoryIndexGenerator } from '../utils/StoryIndexGenerator.ts';
-import type { ChangeDetectionAdapter, FileChangeEvent } from './adapters/index.ts';
-import {
-  ChangeDetectionResolverFactory,
-  DependencyGraphBuilder,
-  IncrementalPatcher,
-  ParseResolveCache,
-} from './dependency-graph/index.ts';
-import type { DependencyGraph, ReverseIndexImpl } from './dependency-graph/index.ts';
+import type { ChangeDetectionAdapter } from './adapters/index.ts';
 import { ChangeDetectionFailureError, ChangeDetectionUnavailableError } from './errors.ts';
 import { GitDiffProvider } from './GitDiffProvider.ts';
 import { extractBaselineEntryIds, IndexBaselineService } from './IndexBaselineService.ts';
-import type { ImportParser } from './parser-registry/index.ts';
-import { ParserRegistry, builtinImportParsers } from './parser-registry/index.ts';
 import { resetChangeDetectionReadiness, setChangeDetectionReadiness } from './readiness.ts';
+import { getStoryIdsByAbsolutePath } from './story-files.ts';
+import { StoryDependencyGraphService } from './StoryDependencyGraphService.ts';
 
 const CHANGE_DETECTION_DEBOUNCE_MS = 200;
 
@@ -47,33 +36,6 @@ function isSameStatus(a: Status | undefined, b: Status): boolean {
     a.sidebarContextMenu === b.sidebarContextMenu &&
     dequal(a.data, b.data)
   );
-}
-
-type StoryIdsByFileCacheKey = Awaited<ReturnType<StoryIndexGenerator['getIndex']>>;
-
-function getStoryIdsByAbsolutePath(
-  cache: WeakMap<
-    StoryIdsByFileCacheKey,
-    { workingDir: string; storyIdsByFile: Map<string, Set<string>> }
-  >,
-  storyIndex: StoryIdsByFileCacheKey,
-  workingDir: string
-): Map<string, Set<string>> {
-  const cached = cache.get(storyIndex);
-  if (cached && cached.workingDir === workingDir) {
-    return cached.storyIdsByFile;
-  }
-  const storyIdsByFile = new Map<string, Set<string>>();
-  Object.values(storyIndex.entries).forEach((entry) => {
-    if (entry.type === 'story' && !entry.importPath.startsWith('virtual:')) {
-      const filePath = normalize(join(workingDir, entry.importPath));
-      const storyIds = storyIdsByFile.get(filePath) ?? new Set<string>();
-      storyIds.add(entry.id);
-      storyIdsByFile.set(filePath, storyIds);
-    }
-  });
-  cache.set(storyIndex, { workingDir, storyIdsByFile });
-  return storyIdsByFile;
 }
 
 export function mergeStatusValues(
@@ -136,10 +98,14 @@ export function buildIndexBaselineStatuses(
 }
 
 /**
- * Coordinates change detection by owning a builder-supplied {@link ChangeDetectionAdapter},
- * eagerly building a reverse-dependency index from story files at startup, applying
- * file-system events incrementally to that index, resolving git-changed files, and publishing
- * the resulting story statuses to the status store.
+ * Publishes change-detection story statuses to the status store. It resolves git-changed files,
+ * maps them to affected stories through a {@link StoryDependencyGraphService} it owns, and emits
+ * `modified`/`affected`/`new` statuses (plus index-baseline `new` entries).
+ *
+ * The module dependency graph itself — building the reverse index, watching builder file events,
+ * and incrementally patching — lives in {@link StoryDependencyGraphService}, which this class
+ * composes. The two responsibilities are independent: the graph never depends on git, the status
+ * store, or the readiness signal, and could be driven by other consumers in the future.
  */
 export class ChangeDetectionService {
   private disposed = false;
@@ -147,29 +113,13 @@ export class ChangeDetectionService {
   private scanInFlight = false;
   private rerunAfterCurrentScan = false;
   private readinessResolved = false;
-  private refreshInFlight = false;
+  private statusPipelineStarted = false;
   private previousStatuses = new Map<string, Status>();
   private gitDiffProvider: GitDiffProvider | undefined;
   private indexBaselineService: IndexBaselineService | undefined;
+  private graph: StoryDependencyGraphService | undefined;
   private readonly workingDir: string;
   private readonly debounceMs: number;
-  private adapter: ChangeDetectionAdapter | undefined;
-  private dependencyGraphBuilder: DependencyGraphBuilder | undefined;
-  private incrementalPatcher: IncrementalPatcher | undefined;
-  private reverseIndex: ReverseIndexImpl | undefined;
-  private storyFiles: Set<string> = new Set();
-  private readonly storyIdsByFileCache = new WeakMap<
-    StoryIdsByFileCacheKey,
-    { workingDir: string; storyIdsByFile: Map<string, Set<string>> }
-  >();
-  /**
-   * Serialises file-change patches so two events touching the same dep set never interleave
-   * across `await` points inside `IncrementalPatcher.patch`. The chain ignores rejections
-   * (each call's failure is logged in {@link handleFileChange}).
-   */
-  private patchQueue: Promise<void> = Promise.resolve();
-  private unsubscribeFileChange: (() => void) | undefined;
-  private unsubscribeStartupFailure: (() => void) | undefined;
 
   constructor(
     private readonly options: {
@@ -179,7 +129,7 @@ export class ChangeDetectionService {
       indexBaselineService?: IndexBaselineService;
       workingDir?: string;
       debounceMs?: number;
-      /** Presets instance used to resolve `experimental_importParsers` contributions from framework/renderer plugins. */
+      /** Presets instance used to resolve `experimental_importParsers` contributions from plugins. */
       presets?: Presets;
     }
   ) {
@@ -210,140 +160,49 @@ export class ChangeDetectionService {
     }
 
     logger.debug('Change detection enabled.');
-    this.adapter = adapter;
 
-    void this.startInternal().catch((error) => {
-      if (this.disposed) {
-        return;
-      }
-      const failure =
-        error instanceof Error ? error : new ChangeDetectionFailureError(String(error));
-      logger.error(`Change detection failed to start: ${failure.message}`);
-      this.resolveReadiness({ status: 'error', error: failure });
-      void this.dispose().catch(() => undefined);
-    });
-  }
-
-  /**
-   * Builds parser registry, resolver, dependency graph, and patcher; subscribes to
-   * file-change events queued behind {@link patchQueue}; kicks off the baseline service
-   * and initial scan.
-   */
-  private async startInternal(): Promise<void> {
-    const adapter = this.adapter;
-    if (!adapter) {
-      return;
-    }
-
-    if (this.disposed) {
-      return;
-    }
-
-    const resolveConfig = await adapter.getResolveConfig();
-    const projectRoot = normalize(resolveConfig.projectRoot ?? this.workingDir);
-
-    const pluginParsers = this.options.presets
-      ? await this.options.presets.apply<ImportParser[]>('experimental_importParsers', [])
-      : [];
-    const registry = new ParserRegistry({
-      defaultParsers: builtinImportParsers,
-      pluginParsers,
-    });
-    const resolver = new ChangeDetectionResolverFactory(resolveConfig);
-    const workspaceRoots = new Set<string>([normalize(getProjectRoot())]);
-
-    const storyIndexGenerator = await this.options.storyIndexGeneratorPromise;
-    const storyIndex = await storyIndexGenerator.getIndex();
-    const storyIdsByFile = getStoryIdsByAbsolutePath(
-      this.storyIdsByFileCache,
-      storyIndex,
-      this.workingDir
-    );
-    this.storyFiles = new Set(storyIdsByFile.keys());
-
-    if (this.disposed) {
-      return;
-    }
-
-    // Shared parse/resolve cache so the patcher reuses cold-start results instead of
-    // re-doing every file's parse + resolution on the first event after boot. The patcher
-    // invalidates per-file entries on every change/unlink before reading.
-    const debugEnv = process.env.STORYBOOK_CHANGE_DETECTION_DEBUG;
-    const cache = new ParseResolveCache({
-      registry,
-      resolver,
-      workspaceRoots,
-      projectRoot,
-      logger,
-      debug: !!debugEnv,
-    });
-
-    this.dependencyGraphBuilder = new DependencyGraphBuilder({
-      registry,
-      resolver,
-      workspaceRoots,
-      projectRoot,
-      cache,
-    });
-
-    // Subscribe BEFORE build — buffer events until patcher is ready
-    const eventBuffer: FileChangeEvent[] = [];
-    this.unsubscribeFileChange = adapter.onFileChange((event) => {
-      if (this.disposed) {
-        return;
-      }
-      eventBuffer.push(event);
-    });
-
-    const { reverseIndex, graph } = await this.dependencyGraphBuilder.build(this.storyFiles);
-    if (this.disposed) {
-      return;
-    }
-    this.reverseIndex = reverseIndex;
-    void this.dumpDebugSnapshot(reverseIndex, graph, projectRoot, workspaceRoots, cache);
-
-    this.incrementalPatcher = new IncrementalPatcher({
-      reverseIndex,
-      graph,
-      registry,
-      resolver,
-      workspaceRoots,
-      projectRoot,
-      cache,
-      isStoryFile: (path: string) => this.storyFiles.has(normalize(path)),
-    });
-
-    // Drain buffered events into patchQueue, then switch to live handler
-    this.unsubscribeFileChange?.();
-    for (const event of eventBuffer) {
-      this.patchQueue = this.patchQueue
-        .then(() => this.handleFileChange(event))
-        .catch(() => undefined);
-    }
-
-    this.unsubscribeFileChange = adapter.onFileChange((event) => {
-      if (this.disposed) {
-        return;
-      }
-      this.patchQueue = this.patchQueue
-        .then(() => this.handleFileChange(event))
-        .catch(() => undefined);
-    });
-
-    if (adapter.onStartupFailure) {
-      this.unsubscribeStartupFailure = adapter.onStartupFailure((event) => {
+    this.graph = new StoryDependencyGraphService({
+      storyIndexGeneratorPromise: this.options.storyIndexGeneratorPromise,
+      workingDir: this.workingDir,
+      presets: this.options.presets,
+      // Status work starts only once the graph is built and queryable.
+      onReady: () => this.startStatusPipeline(),
+      // A file change settled in the graph; recompute statuses (debounced).
+      onChange: () => {
         if (this.disposed) {
           return;
         }
-        logger.warn(`Change detection unavailable: ${event.reason}`);
-        this.resolveReadiness({
-          status: 'unavailable',
-          reason: event.reason,
-          error: event.error,
-        });
+        this.scheduleScan(this.debounceMs);
+      },
+      onError: (failure) => {
+        if (this.disposed) {
+          return;
+        }
+        this.resolveReadiness({ status: 'error', error: failure });
+        void this.dispose().catch(() => undefined);
+      },
+      onUnavailable: (reason, error) => {
+        if (this.disposed) {
+          return;
+        }
+        logger.warn(`Change detection unavailable: ${reason}`);
+        this.resolveReadiness({ status: 'unavailable', reason, error });
         void this.dispose();
-      });
+      },
+    });
+
+    this.graph.start(adapter);
+  }
+
+  /**
+   * Wires the git-diff-driven status pipeline. Runs once the dependency graph is ready (so the
+   * initial scan and every git-state-change scan read a populated reverse index).
+   */
+  private startStatusPipeline(): void {
+    if (this.disposed || this.statusPipelineStarted) {
+      return;
     }
+    this.statusPipelineStarted = true;
 
     void this.getIndexBaselineService().start();
 
@@ -366,130 +225,13 @@ export class ChangeDetectionService {
     if (this.disposed) {
       return;
     }
-    void this.refreshStoryFiles().catch(() => undefined);
-    this.scheduleScan(this.debounceMs);
-  }
-
-  /**
-   * Re-reads the story index and reconciles {@link storyFiles} with stories that have
-   * appeared or disappeared since startup. For each story that newly entered the index, the
-   * patcher is asked to walk it (so its forward edges are recorded). For each story that
-   * left the index, the patcher is asked to unlink it (so its reverse-index entries are
-   * pruned). Replays are queued behind {@link patchQueue} to keep the serialised-patch
-   * invariant intact.
-   */
-  private async refreshStoryFiles(): Promise<void> {
-    if (this.refreshInFlight || !this.incrementalPatcher) {
-      return;
-    }
-    this.refreshInFlight = true;
-    try {
-      const storyIndexGenerator = await this.options.storyIndexGeneratorPromise;
-      const storyIndex = await storyIndexGenerator.getIndex();
-      if (this.disposed) {
-        return;
-      }
-      const storyIdsByFile = getStoryIdsByAbsolutePath(
-        this.storyIdsByFileCache,
-        storyIndex,
-        this.workingDir
-      );
-      const next = new Set(storyIdsByFile.keys());
-      const previous = this.storyFiles;
-
-      const added: string[] = [];
-      for (const path of next) {
-        if (!previous.has(path)) {
-          added.push(path);
-        }
-      }
-      const removed: string[] = [];
-      for (const path of previous) {
-        if (!next.has(path)) {
-          removed.push(path);
-        }
-      }
-
-      if (added.length === 0 && removed.length === 0) {
-        return;
-      }
-
-      this.storyFiles = next;
-
-      for (const path of added) {
-        this.patchQueue = this.patchQueue
-          .then(() => this.handleFileChange({ kind: 'add', path }))
-          .catch(() => undefined);
-      }
-      for (const path of removed) {
-        this.patchQueue = this.patchQueue
-          .then(() => this.handleFileChange({ kind: 'unlink', path }))
-          .catch(() => undefined);
-      }
-    } finally {
-      this.refreshInFlight = false;
-    }
-  }
-
-  private async dumpDebugSnapshot(
-    reverseIndex: ReverseIndexImpl,
-    graph: DependencyGraph,
-    projectRoot: string,
-    workspaceRoots: Set<string>,
-    cache: ParseResolveCache
-  ): Promise<void> {
-    const debugEnv = process.env.STORYBOOK_CHANGE_DETECTION_DEBUG;
-    if (!debugEnv) {
-      return;
-    }
-    const outPath =
-      debugEnv === '1' || debugEnv === 'true'
-        ? join(projectRoot, 'storybook-graph-debug.json')
-        : debugEnv;
-
-    const graphObj: Record<string, string[]> = {};
-    for (const [story, deps] of graph) {
-      graphObj[story] = Array.from(deps).sort();
-    }
-
-    const reverseObj: Record<string, Array<{ story: string; depth: number }>> = {};
-    for (const [dep, stories] of reverseIndex.asMap()) {
-      reverseObj[dep] = Array.from(stories.entries())
-        .map(([story, depth]) => ({ story, depth }))
-        .sort((a, b) => a.depth - b.depth || a.story.localeCompare(b.story));
-    }
-
-    const snapshot = {
-      timestamp: new Date().toISOString(),
-      projectRoot,
-      workspaceRoots: Array.from(workspaceRoots).sort(),
-      // `graph` is keyed by every walked node (story roots + their transitive deps),
-      // and `reverseIndex` records each story root at depth 0 alongside real deps —
-      // so `graph.size` / `reverseIndex.asMap().size` over-report story and dep totals.
-      // Report `storyFiles` from the authoritative source-of-truth set, plus the raw
-      // node/entry counts under unambiguous names for diagnostics.
-      storyFiles: this.storyFiles.size,
-      graphNodes: graph.size,
-      reverseIndexEntries: reverseIndex.asMap().size,
-      graph: graphObj,
-      reverseIndex: reverseObj,
-      // Each entry records one named-import barrel lookup: which names were requested,
-      // which source files they resolved to, and whether the barrel itself was also
-      // included (needBarrel: true means at least one name fell back to the barrel).
-      barrelResolutions: cache.getBarrelTrace() ?? [],
-    };
-
-    try {
-      await writeFile(outPath, JSON.stringify(snapshot, null, 2), 'utf8');
-      logger.debug(`Change detection: graph debug snapshot written to ${outPath}`);
-    } catch (error) {
-      logger.warn(
-        `Change detection: failed to write debug snapshot to ${outPath}: ${error instanceof Error ? error.message : String(error)}`
-      );
-    }
+    this.graph?.onStoryIndexInvalidated();
   }
 
   async dispose(): Promise<void> {
+    if (this.disposed) {
+      return;
+    }
     this.disposed = true;
     this.rerunAfterCurrentScan = false;
 
@@ -498,33 +240,9 @@ export class ChangeDetectionService {
       this.debounceTimer = undefined;
     }
 
-    this.unsubscribeFileChange?.();
-    this.unsubscribeFileChange = undefined;
-    this.unsubscribeStartupFailure?.();
-    this.unsubscribeStartupFailure = undefined;
-
     this.gitDiffProvider?.dispose();
-    // Drain in-flight patches before tearing down the OXC parse pool so no
-    // patch reads the pool after it has been disposed.
-    await this.patchQueue.catch(() => undefined);
-    await disposeOxcParsePool().catch(() => undefined);
-  }
-
-  private async handleFileChange(event: FileChangeEvent): Promise<void> {
-    if (this.disposed || !this.incrementalPatcher) {
-      return;
-    }
-    try {
-      await this.incrementalPatcher.patch(event);
-    } catch (error) {
-      logger.warn(
-        `Change detection: failed to apply ${event.kind} for ${event.path}: ${error instanceof Error ? error.message : String(error)}`
-      );
-    }
-    if (this.disposed) {
-      return;
-    }
-    this.scheduleScan(this.debounceMs);
+    // Tear down the graph last: it drains in-flight patches before disposing the OXC parse pool.
+    await this.graph?.dispose();
   }
 
   private scheduleScan(delayMs: number): void {
@@ -539,17 +257,16 @@ export class ChangeDetectionService {
   }
 
   private async scan(): Promise<void> {
-    if (this.disposed || !this.reverseIndex) {
+    if (this.disposed || !this.graph) {
       return;
     }
 
-    // Snapshot and drain the current patch chain before reading reverseIndex. Without this
-    // await, a scan triggered mid-patch (between removeStory and the re-walk's recordEdges)
-    // reads a transiently empty reverseIndex and publishes incorrect statuses.
-    const patchSnapshot = this.patchQueue;
-    await patchSnapshot.catch(() => undefined);
+    // Drain the graph's patch chain before reading it. Without this, a scan triggered mid-patch
+    // (between a story's removeStory and the re-walk's recordEdges) reads a transiently empty
+    // reverse index and publishes incorrect statuses.
+    await this.graph.whenSettled();
 
-    if (this.disposed || !this.reverseIndex) {
+    if (this.disposed || !this.graph.hasGraph()) {
       return;
     }
 
@@ -561,7 +278,7 @@ export class ChangeDetectionService {
     this.scanInFlight = true;
 
     try {
-      const nextStatuses = await this.buildStatuses(this.reverseIndex);
+      const nextStatuses = await this.buildStatuses();
       if (this.disposed) {
         return;
       }
@@ -608,7 +325,7 @@ export class ChangeDetectionService {
     }
   }
 
-  private async buildStatuses(reverseIndex: ReverseIndexImpl): Promise<Map<string, Status>> {
+  private async buildStatuses(): Promise<Map<string, Status>> {
     const gitDiffProvider = this.getGitDiffProvider();
     const [changes, repoRoot, storyIndexGenerator, baselineEntryIds] = await Promise.all([
       gitDiffProvider.getChangedFiles(),
@@ -627,15 +344,11 @@ export class ChangeDetectionService {
 
     const storyIndex = await storyIndexGenerator.getIndex();
     const baselineStatuses = buildIndexBaselineStatuses(storyIndex, baselineEntryIds);
-    const storyIdsByFile = getStoryIdsByAbsolutePath(
-      this.storyIdsByFileCache,
-      storyIndex,
-      this.workingDir
-    );
+    const storyIdsByFile = getStoryIdsByAbsolutePath(storyIndex, this.workingDir);
     const statuses = new Map<string, Status>();
 
     for (const changedFile of scannedFiles) {
-      const affectedStoryFiles = reverseIndex.lookup(changedFile);
+      const affectedStoryFiles = this.graph?.lookup(changedFile) ?? new Map<string, number>();
       // Include the changed file as a story-at-distance-0 if it IS a story (parity with
       // legacy trace-changed.ts:10-12).
       const allEntries = new Map(affectedStoryFiles);
