+++ code/core/src/core-server/change-detection/change-detection.test-helpers.ts
@@ -1,0 +1,141 @@
+import { vi } from 'vitest';
+
+import type { StoryIndex } from 'storybook/internal/types';
+
+import type { ChangeDetectionAdapter, FileChangeEvent } from './adapters/index.ts';
+import {
+  ChangeDetectionResolverFactory,
+  DependencyGraphBuilder,
+  IncrementalPatcher,
+  ReverseIndexImpl,
+} from './dependency-graph/index.ts';
+
+/**
+ * Shared scaffolding for the change-detection unit tests. The dependency-graph constructors are
+ * mocked per test file (each file declares its own `vi.mock('./dependency-graph/index.ts', ...)`);
+ * these helpers drive those mocks and build synthetic adapters / indexes / reverse indexes.
+ */
+
+export function createDeferred<T>() {
+  let resolve!: (value: T) => void;
+
+  return {
+    promise: new Promise<T>((fulfill) => {
+      resolve = fulfill;
+    }),
+    resolve,
+  };
+}
+
+export function createStoryIndex(
+  entries: Array<{ storyId: string; importPath: string; title?: string; name?: string }>
+): StoryIndex {
+  return {
+    v: 5,
+    entries: Object.fromEntries(
+      entries.map(({ storyId, importPath, title = 'Story', name = 'Default' }) => [
+        storyId,
+        {
+          id: storyId,
+          type: 'story',
+          subtype: 'story',
+          title,
+          name,
+          importPath,
+        },
+      ])
+    ),
+  };
+}
+
+export interface MockAdapterHandle {
+  adapter: ChangeDetectionAdapter;
+  emitFileChange: (event: FileChangeEvent) => void;
+  emitStartupFailure: (event: { reason: string; error?: Error }) => void;
+  hasFileChangeSubscriber: () => boolean;
+  hasStartupFailureSubscriber: () => boolean;
+}
+
+export function createMockAdapter(opts?: {
+  resolveConfig?: { projectRoot?: string };
+  withoutStartupFailure?: boolean;
+}): MockAdapterHandle {
+  const fileHandlers = new Set<(e: FileChangeEvent) => void>();
+  const startupHandlers = new Set<(e: { reason: string; error?: Error }) => void>();
+
+  const adapter: ChangeDetectionAdapter = {
+    async getResolveConfig() {
+      return {
+        projectRoot: opts?.resolveConfig?.projectRoot ?? '/repo',
+      };
+    },
+    onFileChange(handler) {
+      fileHandlers.add(handler);
+      return () => fileHandlers.delete(handler);
+    },
+  };
+
+  if (!opts?.withoutStartupFailure) {
+    adapter.onStartupFailure = (handler) => {
+      startupHandlers.add(handler);
+      return () => startupHandlers.delete(handler);
+    };
+  }
+
+  return {
+    adapter,
+    emitFileChange: (event) => {
+      fileHandlers.forEach((h) => h(event));
+    },
+    emitStartupFailure: (event) => {
+      startupHandlers.forEach((h) => h(event));
+    },
+    hasFileChangeSubscriber: () => fileHandlers.size > 0,
+    hasStartupFailureSubscriber: () => startupHandlers.size > 0,
+  };
+}
+
+/**
+ * Build a ReverseIndexImpl populated with the given (dep -> story -> depth) entries.
+ * Used by tests to control what `reverseIndex.lookup(changedFile)` returns.
+ */
+export function buildReverseIndex(
+  edges: Iterable<readonly [string, string, number]>
+): ReverseIndexImpl {
+  const reverseIndex = new ReverseIndexImpl();
+  for (const [dep, story, depth] of edges) {
+    reverseIndex.record(dep, story, depth);
+  }
+  return reverseIndex;
+}
+
+/**
+ * Stub the dependency-graph constructors so the service under test uses an in-test
+ * ReverseIndexImpl + an inert IncrementalPatcher.
+ *
+ * Note: `vi.mock` replaces these exports with plain `vi.fn()` constructors. When the service calls
+ * `new Ctor(...)` we must return objects via `mockImplementation` — but vitest invokes the impl
+ * with `Reflect.construct` on `new`, so arrow-function impls throw "is not a constructor".
+ * `function () { return obj; }` works because regular functions support `[[Construct]]`.
+ */
+export function installDependencyGraphMocks(reverseIndex: ReverseIndexImpl): {
+  patchSpy: ReturnType<typeof vi.fn>;
+  buildSpy: ReturnType<typeof vi.fn>;
+} {
+  const patchSpy = vi.fn(async () => undefined);
+  const buildSpy = vi.fn(async () => ({ reverseIndex, graph: new Map() }));
+
+  vi.mocked(ChangeDetectionResolverFactory).mockImplementation(function () {
+    return {
+      resolve: vi.fn(async () => null),
+    } as unknown as ChangeDetectionResolverFactory;
+  } as unknown as new () => ChangeDetectionResolverFactory);
+  vi.mocked(DependencyGraphBuilder).mockImplementation(function () {
+    return { build: buildSpy } as unknown as DependencyGraphBuilder;
+  } as unknown as new () => DependencyGraphBuilder);
+  vi.mocked(IncrementalPatcher).mockImplementation(function () {
+    return { patch: patchSpy } as unknown as IncrementalPatcher;
+  } as unknown as new () => IncrementalPatcher);
+
+  return { patchSpy, buildSpy };
+}
